refactor admin api auth
